Modified methods of Perceptrons to detect a line with 0<x<1 and 0<y<1

+++ Excercises/src/Perceptron/AbstractPerceptron.java
@@ -1,13 +1,13 @@
 package Perceptron;
 
-public abstract class AbstractPerceptron implements Perceptron{
-    protected int w1;
-    protected int w2;
-    protected int b;
+public abstract class AbstractPerceptron implements Perceptron {
+    protected double w1;
+    protected double w2;
+    protected double b;
     private double lr;
 
 
-    AbstractPerceptron(int w1,int w2,int b){
+    AbstractPerceptron(int w1, int w2, int b) {
         this.w1 = w1;
         this.w2 = w2;
         this.b = b;
@@ -16,37 +16,35 @@ public abstract class AbstractPerceptron implements Perceptron{
     }
 
     @Override
-    public int check(int x1, int x2) {
+    public int check(double x1, double x2) {
         // if weighted sum + b <= 0 -> false
-        if(x1*w1 + x2*w2 + b <= 0){
-            return 0;}
-        else{
+        if (x1 * w1 + x2 * w2 + b <= 0) {
+            return 0;
+        } else {
             return 1;
         }
 
     }
 
-
-
-    public int getW1() {
+    public double getW1() {
         return w1;
     }
 
-    public int getW2() {
+    public double getW2() {
         return w2;
     }
 
-    public int getB() {
+    public double getB() {
         return b;
     }
 
-    public void learn(int x1, int x2,int desiredOutput){
-        int realOutput = check(x1,x2);
+    public void learn(double x1, double x2, int desiredOutput) {
+        int realOutput = check(x1, x2);
         int diff = desiredOutput - realOutput;
 
-        this.w1+= lr*x1*diff;
-        this.w2+= lr*x2*diff;
+        this.w1 += lr * x1 * diff;
+        this.w2 += lr * x2 * diff;
 
-        b+= lr*diff;
+        b += lr * diff;
     }
 }
